Extract Manifest logic into individual projects
The Manifest was left in the parent POM when the second module was created. This led to both modules having the same WatPluginMainClass entry in the manifest, which is not correct. This relocates the manifest code into the individual POMs, and sets the correct (different) values in the correct jars.